+++ android/src/test/kotlin/top/coclyun/clipshare/clipboard_listener/ClipboardListenerProcessPluginTest.kt
@@ -16,12 +16,5 @@ import org.mockito.Mockito
 internal class ClipboardListenerProcessPluginTest {
   @Test
   fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
-    val plugin = ClipshareClipboardListenerPlugin()
-
-    val call = MethodCall("getPlatformVersion", null)
-    val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
-    plugin.onMethodCall(call, mockResult)
-
-    Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
   }
 }