+++ lib/enums.dart
@@ -10,6 +10,7 @@ enum ClipboardContentType {
         (e) => e.name.toUpperCase() == value.toUpperCase(),
         orElse: () {
           debugPrint("ClipboardContentType '$value' unknown");
+          throw 'bbb';
           return ClipboardContentType.unknown;
         },
       );